+++ ansible/Vagrantfile
@@ -1,4 +1,4 @@
-ENV['VAGRANT_DEFAULT_PROVIDER'] = 'virtualbox'
+#ENV['VAGRANT_DEFAULT_PROVIDER'] = 'virtualbox'
 Vagrant.configure("2") do |config|
 
   config.vm.provider :virtualbox do |v|
