+++ ansible/roles/db/molecule/default/molecule.yml
@@ -1,6 +1,6 @@
 ---
-dependency:
-  name: galaxy
+#dependency:
+#  name: galaxy
 driver:
   name: vagrant
   provider:
@@ -14,9 +14,9 @@ provisioner:
   name: ansible
   lint:
     name: ansible-lint
-scenario:
-  name: default
-verifier:
-  name: testinfra
-  lint:
-    name: flake8
+#scenario:
+#  name: default
+#verifier:
+#  name: testinfra
+#  lint:
+#    name: flake8
