+++ ansible/roles/db/tasks/config_mongo.yml
@@ -4,3 +4,4 @@
     dest: /etc/mongod.conf
     mode: 0644
   notify: restart mongod
+
